added arguments

+++ destroy.sh
@@ -1,5 +1,12 @@
 #!/bin/sh
 
+die () {
+   echo >&2 "USAGE: destroy.sh <container_name>"
+   exit 1
+}
+
+[ "$#" -eq 1 ] || die 
+
 CONTAINER_NAME=nagiost1
 lxc-stop -n ${CONTAINER_NAME}
 lxc-destroy -n ${CONTAINER_NAME}
